move completion code to completion provider
